All primary functions implemented
Profiles can be created and deleted and contain avatar, background, name, head data, torso data, and leg data. all sizing data can be edited and is saved with each profile. Profiles can be swapped to view. All that needs to be done is make the profiles save when program stops.